requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
discord.py==2.4.0
frozenlist==1.5.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
httplib2==0.22.0
idna==3.10
multidict==6.1.0
oauth2client==4.1.3
oauthlib==3.2.2
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
urllib3==2.2.3
yarl==1.17.1